+++ @xen-orchestra/lite/src/components/vm/VmHeader.vue
@@ -2,14 +2,19 @@
   <TitleBar :icon="faDisplay">
     {{ name }}
     <template #actions>
-      <AppMenu shadow>
+      <AppMenu shadow placement="bottom-end">
         <template #trigger="{ open, isOpen }">
           <UiButton :active="isOpen" :icon="faPowerOff" @click="open">
             {{ $t("change-state") }}
             <UiIcon :icon="faAngleDown" />
           </UiButton>
         </template>
-        <MenuItem :disabled="!isHalted" :icon="faPlay">
+        <MenuItem
+          @click="start"
+          :busy="isBusy('start')"
+          :disabled="!isHalted"
+          :icon="faPlay"
+        >
           {{ $t("start") }}
         </MenuItem>
         <MenuItem :disabled="!isHalted" :icon="faServer">
@@ -17,6 +22,8 @@
           <template #submenu>
             <MenuItem
               v-for="host in hostStore.allRecords"
+              @click="startOn(host.$ref)"
+              :busy="isBusy('start_on')"
               v-bind:key="host.$ref"
               :icon="faServer"
             >
@@ -37,26 +44,58 @@
             </MenuItem>
           </template>
         </MenuItem>
-        <MenuItem :disabled="!isRunning" :icon="faPause">
+        <MenuItem
+          @click="pause"
+          :busy="isBusy('pause')"
+          :disabled="!isRunning"
+          :icon="faPause"
+        >
           {{ $t("pause") }}
         </MenuItem>
-        <MenuItem :disabled="!isRunning" :icon="faMoon">
+        <MenuItem
+          @click="suspend"
+          :busy="isBusy('suspend')"
+          :disabled="!isRunning"
+          :icon="faMoon"
+        >
           {{ $t("suspend") }}
         </MenuItem>
         <!-- TODO: update the icon once Clémence has integrated the action into figma -->
-        <MenuItem :disabled="!isSuspended && !isPaused" :icon="faCirclePlay">
+        <MenuItem
+          @click="resume"
+          :busy="isBusy('unpause') || isBusy('resume')"
+          :disabled="!isSuspended && !isPaused"
+          :icon="faCirclePlay"
+        >
           {{ $t("resume") }}
         </MenuItem>
-        <MenuItem :disabled="!isRunning" :icon="faRotateLeft">
+        <MenuItem
+          @click="reboot"
+          :busy="isBusy('clean_reboot')"
+          :disabled="!isRunning"
+          :icon="faRotateLeft"
+        >
           {{ $t("reboot") }}
         </MenuItem>
-        <MenuItem :disabled="!isRunning && !isPaused" :icon="faRepeat">
+        <MenuItem
+          @click="reboot(true)"
+          :busy="isBusy('hard_reboot')"
+          :disabled="!isRunning && !isPaused"
+          :icon="faRepeat"
+        >
           {{ $t("force-reboot") }}
         </MenuItem>
-        <MenuItem :disabled="!isRunning" :icon="faPowerOff">
+        <MenuItem
+          @click="shutdown"
+          :busy="isBusy('clean_shutdown')"
+          :disabled="!isRunning"
+          :icon="faPowerOff"
+        >
           {{ $t("shutdown") }}
         </MenuItem>
         <MenuItem
+          @click="shutdown(true)"
+          :busy="isBusy('hard_shutdown')"
           :disabled="!isRunning && !isSuspended && !isPaused"
           :icon="faPlug"
         >
@@ -99,14 +138,27 @@ const hostStore = useHostStore();
 const poolStore = usePoolStore();
 const { currentRoute } = useRouter();
 
-const vm = computed(() =>
-  vmStore.getRecordByUuid(currentRoute.value.params.uuid as string)
+const isBusy = (operation: string) =>
+  Object.values(vm.value.current_operations).includes(operation);
+const vm = computed(
+  () => vmStore.getRecordByUuid(currentRoute.value.params.uuid as string)!
 );
-const name = computed(() => vm.value?.name_label);
-const isRunning = computed(() => vm.value?.power_state === "Running");
-const isHalted = computed(() => vm.value?.power_state === "Halted");
-const isSuspended = computed(() => vm.value?.power_state === "Suspended");
-const isPaused = computed(() => vm.value?.power_state === "Paused");
+const name = computed(() => vm.value.name_label);
+const isRunning = computed(() => vm.value.power_state === "Running");
+const isHalted = computed(() => vm.value.power_state === "Halted");
+const isSuspended = computed(() => vm.value.power_state === "Suspended");
+const isPaused = computed(
+  () =>
+    vm.value.power_state === "Paused" &&
+    !(isBusy("clean_shutdown") || isBusy("hard_shutdown"))
+);
+const start = () => vmStore.start([vm.value.$ref]);
+const startOn = (hostRef: string) => vmStore.startOn([vm.value.$ref], hostRef);
+const pause = () => vmStore.pause([vm.value.$ref]);
+const suspend = () => vmStore.suspend([vm.value.$ref]);
+const resume = () => vmStore.resume([[vm.value.$ref, vm.value.power_state]]);
+const reboot = (force?: boolean) => vmStore.reboot([vm.value.$ref], force);
+const shutdown = (force?: boolean) => vmStore.shutdown([vm.value.$ref], force);
 </script>
 
 <style lang="postcss" scoped>
